Fixed two-part dependent-vowel conversion issue
The word document splits paragraphs in to runs and each run has unique
font and style. Unfortunately in some document the split is done in
cruel way, that is "Nrh" = "சோ" where the split happens between Nr and h,
and as a result the conversion to Unicode gets flawed.

This is fixed by changing the algorithm of grouping texts of similar
font and then converting them as a whole. Hence the cruel split is being
overcome!